Drop support for python 3.9
Python 3.9 in combination with Pyside6 asyncio results in the following error:
AttributeError: 'QAsyncioTask' object has no attribute '_source_traceback'
According to https://bugreports.qt.io/projects/PYSIDE/issues/PYSIDE-2972?filter=allissues it seems that this is from version 6.7 and up
so it seems that it won't be fixed anymore.
Also, Python 3.9 is almost end-of-life, and the latest pygfx has also
dropped support for 3.9.