Fine-grained reactivity (#119)
This PR completely does away with the React way of having a virtual DOM and comparing trees for figuring out which changes are needed. In its place there is now a 'fine-grained reactivity' system that updates just what is needed and does away with the overhead of having to compare virtual DOMs. This results in improved performance in scenarios with lots of moving parts¹.

Notable API changes:

the create_element (h) methods are gone: currently only possible to create components with template syntax. This might be coming back at a later stage
signature of Collagraph.render method changed:
takes a Component class instead of a virtual VNode
added state that gets passed to the component (created from the given class)
removed callback method in favor of subscribing to changes directly from the renderer
text elements (moustaches) are not supported (yet)
Changes to the template syntax:

No need to have a single root item to render in a component
No need for enclosing template tag: all tags outside the script tag are rendered
These changes should make the template syntax a little more concise (thanks to Svelte for the inspiration).

Other notable changes:

* Some small fixes for PySide6 updates
* Bump version to 0.8.0



¹ There is no official benchmark to prove or quantify the performance improvements